finance-rag-api/requirements.txt
# LLM APIs
openai>=1.0.0
anthropic>=0.18.0

# RAG Components
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development
pytest>=8.0.0
httpx>=0.26.0
jupyter>=1.0.0

# Monitoring (Production)
prometheus-client>=0.19.0

# Web Demo
streamlit>=1.30.0
